requirements.txt
streamlit
python-dotenv
langchain_huggingface
langchain_groq
langchain
langchain_text_splitters
langchain_core
langchain_chroma
langchain_community
ipykernel
pypdf
lxml
bs4
arxiv
pymupdf
wikipdia    
chromeadb
sentence_transformers
faiss-cpu
langserve
